@@ -4,7 +4,6 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "3.5.6",
-    "@binance/w3w-utils": "1.0.1-alpha.0",
     "@davatar/react": "1.8.1",
     "@ensdomains/ens-avatar": "1.0.0-alpha.0.ethers.6",
     "@ethersproject/providers": "5.5.1",
@@ -14,18 +13,21 @@
     "@headlessui/react": "1.6.1",
     "@lingui/core": "3.13.3",
     "@lingui/react": "3.13.3",
-    "@rainbow-me/rainbowkit": "0.12.18",
+    "@rainbow-me/rainbowkit": "2.0.1",
+    "@tanstack/react-query": "5.25.0",
     "@taskworld.com/rereselect": "0.3.0",
     "@testing-library/jest-dom": "5.16.1",
     "@testing-library/react": "11.2.7",
     "@testing-library/user-event": "12.8.3",
     "@types/lodash": "4.14.198",
     "@types/node": "18.7.13",
-    "@types/react": "18.0.17",
-    "@types/react-dom": "18.0.6",
+    "@types/react": "18.2.64",
+    "@types/react-dom": "18.2.21",
     "@types/react-router-dom": "5.3.3",
     "@uniswap/sdk-core": "3.0.1",
     "@uniswap/v3-sdk": "3.9.0",
+    "@wagmi/connectors": "4.1.14",
+    "@wagmi/core": "2.6.5",
     "bigdecimal": "0.6.1",
     "classnames": "2.3.1",
     "crypto-js": "4.2.0",
@@ -43,9 +45,9 @@
     "qrcode.react": "^3.1.0",
     "query-string": "7.1.1",
     "rc-slider": "9.7.5",
-    "react": "17.0.2",
     "react-date-range": "2.0.0-alpha.4",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-helmet": "6.1.0",
     "react-hot-toast": "2.2.0",
     "react-icons": "4.3.1",
@@ -63,10 +65,10 @@
     "sass": "1.55.0",
     "shallowequal": "1.1.0",
     "swr": "2.2.1",
-    "typescript": "4.7.4",
+    "typescript": "5.4.2",
     "use-context-selector": "1.4.1",
-    "viem": "1.1.4",
-    "wagmi": "0.12.19",
+    "viem": "2.7.20",
+    "wagmi": "2.5.7",
     "web-vitals": "1.1.2"
   },
   "resolutions": {
@@ -80,8 +82,8 @@
     "build": "INLINE_RUNTIME_CHUNK=false react-app-rewired build",
     "build-home": "INLINE_RUNTIME_CHUNK=false env-cmd -e production-home react-app-rewired build",
     "build-app": "INLINE_RUNTIME_CHUNK=false env-cmd -e production-app react-app-rewired build",
-    "test": "react-app-rewired test --transformIgnorePatterns \"node_modules/(?!@wagmi)/\"",
-    "test:ci": "react-app-rewired test --watchAll=false --transformIgnorePatterns \"node_modules/(?!@wagmi)/\"",
+    "test": "react-app-rewired test",
+    "test:ci": "react-app-rewired test --watchAll=false",
     "eject": "react-scripts eject",
     "lint": "node node_modules/.bin/eslint src",
     "pre-commit": "npm run test -- --watchAll=false && npm run lint",
