@@ -1,26 +1,25 @@
-import React from "react";
-import ReactDOM from "react-dom";
 import "regenerator-runtime/runtime";
+
+import React from "react";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
+
+import WalletProvider from "lib/wallets/WalletProvider";
 import App from "./App/App";
 import reportWebVitals from "./reportWebVitals";
-import WalletProvider from "lib/wallets/WalletProvider";
 
-ReactDOM.render(
+createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <Router>
       <WalletProvider>
         <App />
       </WalletProvider>
     </Router>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.info))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-export { formatUsd } from "./lib/numbers";
-export { formatTokenAmountWithUsd } from "./lib/numbers";
-export { formatTokenAmount } from "./lib/numbers";
+export { formatTokenAmount, formatTokenAmountWithUsd, formatUsd } from "./lib/numbers";
