@@ -1,6 +1,6 @@
 const webpack = require("webpack");
 
-module.exports = function override(config) {
+function override(config) {
   // https://github.com/lingui/js-lingui/issues/1195
   // Adding loader to use for .po files to webpack
   const loaders = config.module.rules.find((rule) => Array.isArray(rule.oneOf));
@@ -25,5 +25,18 @@ module.exports = function override(config) {
       Buffer: ["buffer", "Buffer"],
     }),
   ]);
+
+  return config;
+}
+
+override.jest = function (config) {
+  config.transform = {
+    "node_modules/multiformats/(basics|block|cid)$": "react-app-rewired/scripts/utils/babelTransform",
+    ...config.transform,
+  };
+  config.transformIgnorePatterns = ["node_modules/(?!multiformats|wagmi)/"];
+
   return config;
 };
+
+module.exports = override;
