@@ -1,30 +1,10 @@
-import React from "react";
+import { darkTheme, RainbowKitProvider, Theme } from "@rainbow-me/rainbowkit";
 import "@rainbow-me/rainbowkit/styles.css";
-import { connectorsForWallets, darkTheme, RainbowKitProvider, Theme, WalletList } from "@rainbow-me/rainbowkit";
-import {
-  ledgerWallet,
-  safeWallet,
-  rabbyWallet,
-  injectedWallet,
-  metaMaskWallet,
-  walletConnectWallet,
-  coinbaseWallet,
-  rainbowWallet,
-  imTokenWallet,
-  zerionWallet,
-  okxWallet,
-} from "@rainbow-me/rainbowkit/wallets";
-import { configureChains, createClient, WagmiConfig } from "wagmi";
-import { arbitrum, arbitrumGoerli, avalanche, avalancheFuji } from "wagmi/chains";
-import { publicProvider } from "wagmi/providers/public";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import merge from "lodash/merge";
-import { isDevelopment } from "config/env";
-import { coreWallet } from "./connecters/core/coreWallet";
-import { bitgetWallet } from "./connecters/bitgetWallet/bitgetWallet";
-import binanceWallet from "./connecters/binanceW3W/binanceWallet";
+import { WagmiProvider } from "wagmi";
 
-const WALLET_CONNECT_PROJECT_ID = "de24cddbaf2a68f027eae30d9bb5df58";
-const APP_NAME = "GMX";
+import { rainbowKitConfig } from "./rainbowKitConfig";
 
 const walletTheme = merge(darkTheme(), {
   colors: {
@@ -38,55 +18,16 @@ const walletTheme = merge(darkTheme(), {
   },
 } as Theme);
 
-const { chains, provider } = configureChains(
-  [arbitrum, avalanche, ...(isDevelopment() ? [arbitrumGoerli, avalancheFuji] : [])],
-  [publicProvider()]
-);
-
-const recommendedWalletList: WalletList = [
-  {
-    groupName: "Recommended",
-    wallets: [
-      injectedWallet({ chains }),
-      safeWallet({ chains }),
-      rabbyWallet({ chains }),
-      metaMaskWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      walletConnectWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-    ],
-  },
-];
-
-const othersWalletList: WalletList = [
-  {
-    groupName: "Others",
-    wallets: [
-      coreWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      coinbaseWallet({ appName: APP_NAME, chains }),
-      binanceWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      okxWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      ledgerWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      rainbowWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      bitgetWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      zerionWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-      imTokenWallet({ chains, projectId: WALLET_CONNECT_PROJECT_ID }),
-    ],
-  },
-];
-
-const connectors = connectorsForWallets([...recommendedWalletList, ...othersWalletList]);
-
-const wagmiClient = createClient({
-  autoConnect: true,
-  connectors,
-  provider,
-});
+const queryClient = new QueryClient();
 
 export default function WalletProvider({ children }) {
   return (
-    <WagmiConfig client={wagmiClient}>
-      <RainbowKitProvider theme={walletTheme} chains={chains} modalSize="compact">
-        {children}
-      </RainbowKitProvider>
-    </WagmiConfig>
+    <WagmiProvider config={rainbowKitConfig}>
+      <QueryClientProvider client={queryClient}>
+        <RainbowKitProvider theme={walletTheme} modalSize="compact">
+          {children}
+        </RainbowKitProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
   );
 }
