@@ -1,5 +1,6 @@
-import { switchNetwork as switchNetworkWagmi } from "@wagmi/core";
+import { switchChain } from "@wagmi/core";
 import { SELECTED_NETWORK_LOCAL_STORAGE_KEY } from "config/localStorage";
+import { rainbowKitConfig } from "./rainbowKitConfig";
 
 export type NetworkMetadata = {
   chainId: string;
@@ -15,7 +16,9 @@ export type NetworkMetadata = {
 
 export async function switchNetwork(chainId, active) {
   if (active) {
-    await switchNetworkWagmi({ chainId });
+    await switchChain(rainbowKitConfig, {
+      chainId,
+    });
   } else {
     // chainId in localStorage allows to switch network even if wallet is not connected
     // or there is no wallet at all
